backend/requirements.txt
# Core AI Provider Dependencies
anthropic>=0.7.0
google-generativeai>=0.3.0

# Data Processing
pandas>=1.5.0
openpyxl>=3.0.0

# Environment and Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0
pydantic>=2.0.0

# Utilities
requests>=2.28.0
typing-extensions>=4.0.0

# Web Framework
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3

# Voice Processing Dependencies
openai-whisper>=20231117
pyaudio>=0.2.11
numpy>=1.21.0
torch>=1.9.0